+++ src/hanami/src/core/cluster/cluster_init.cpp
@@ -120,7 +120,6 @@ initializeInputs(Cluster* cluster, const ClusterMeta& clusterMeta)
 
         Hexagon* targetHexagon = &cluster->hexagons[inputInterface.targetHexagonId];
         targetHexagon->header.isInputHexagon = true;
-        targetHexagon->header.isBinaryInput = inputMeta.binary;
         targetHexagon->inputInterface = &cluster->inputInterfaces[inputMeta.name];
     }
 }