+++ src/libraries/hanami_cluster_parser/include/hanami_cluster_parser/cluster_meta.h
@@ -38,7 +38,6 @@ namespace Hanami
 struct InputMeta {
     std::string name = "";
     uint32_t targetHexagonId = UNINTI_POINT_32;
-    bool binary = false;
 };
 
 struct OutputMeta {